[Change] Signup내 Ratio, BodyFigure 컴포넌트 분리
