+++ src/pages/Signup/SignupBodyFigure/components/Ratio/Ratio.jsx
@@ -1,6 +1,6 @@
 import * as S from "./StyledRatio"
-import plus from "../../../../assets/images/plus.png"
-import minus from "../../../../assets/images/minus.png"
+import plus from "../../../../../assets/images/plus.png"
+import minus from "../../../../../assets/images/minus.png"
 const Ratio = ({ratioValue, ratioText, handleRatio}) => {
 	return (
 		<S.RatioWrapper>
