+++ src/pages/Signup/SignupBodyFigure/components/Ratio/StyledRatio.js
@@ -1,5 +1,5 @@
 import {styled} from "styled-components"
-import theme from "../../../../styles/theme"
+import theme from "../../../../../styles/theme"
 
 const RatioButton = styled.button`
 	border-radius: 34px;
